feat: fetch slack data from db
