+++ tests/Unit/Authenticator/SignatureAuthenticatorTest.php
