+++ tests/Unit/PullRequest/ClosePrUseCaseTest.php
@@ -2,40 +2,36 @@
 
 declare(strict_types=1);
 
-namespace App\Tests\App\Tests\Unit\PullRequest;
+namespace App\Tests\Unit\PullRequest;
 
 use App\Entity\SlackMessage;
-use App\PullRequest\DraftPrUseCase;
+use App\PullRequest\ClosePrUseCase;
 use App\Repository\SlackMessageRepositoryInterface;
 use App\Slack\SlackMessengerInterface;
 use App\Transfers\WebHookTransfer;
-use Doctrine\ORM\EntityManagerInterface;
 use PHPUnit\Framework\TestCase;
 use Psr\Log\LoggerInterface;
 
-class DraftPrUseCaseTest extends TestCase
+class ClosePrUseCaseTest extends TestCase
 {
-    private EntityManagerInterface $entityManager;
-
     private SlackMessageRepositoryInterface $slackMessageRepository;
 
     private SlackMessengerInterface $slackMessenger;
 
-    private DraftPrUseCase $useCase;
+    private ClosePrUseCase $useCase;
 
     protected function setUp(): void
     {
-        $this->entityManager = $this->createMock(EntityManagerInterface::class);
         $this->slackMessageRepository = $this->createMock(SlackMessageRepositoryInterface::class);
         $this->slackMessenger = $this->createMock(SlackMessengerInterface::class);
-        $this->useCase = new DraftPrUseCase(
-            $this->entityManager,
+        $this->useCase = new ClosePrUseCase(
             $this->slackMessageRepository,
             $this->slackMessenger,
             $this->createMock(LoggerInterface::class)
         );
     }
-    public function testHandleDraftFromScratch(): void
+
+    public function testHandleDoesNothingIfNoMessageStoredForPrNumber(): void
     {
         // Arrange
         $webHookTransfer = new WebHookTransfer(
@@ -51,15 +47,14 @@ class DraftPrUseCaseTest extends TestCase
             ->willReturn(null);
 
         $this->slackMessenger->expects($this->never())
-            ->method('sendNewMessage');
-        $this->entityManager->expects($this->never())
-            ->method('persist');
+            ->method('updateMessage');
 
         // Act
         $this->useCase->handle($webHookTransfer);
     }
 
-    public function testHandleDraftPrAfterReadyForReview(): void
+
+    public function testHandleSendsMessage(): void
     {
         // Arrange
         $webHookTransfer = new WebHookTransfer(
@@ -68,22 +63,15 @@ class DraftPrUseCaseTest extends TestCase
             'testuser'
         );
 
-        $slackMessageEntity = new SlackMessage()->setPrNumber(42)->setTs('12345.6789');
-
         // Assert
         $this->slackMessageRepository->expects($this->once())
             ->method('findOneByPrNumber')
             ->with(42)
-            ->willReturn($slackMessageEntity);
+            ->willReturn(new SlackMessage());
 
         $this->slackMessenger->expects($this->once())
-            ->method('removeMessage')
-            ->with($slackMessageEntity)
-            ->willReturn(true);
+            ->method('updateMessage');
 
-        $this->entityManager->expects($this->once())
-            ->method('remove')
-            ->with($slackMessageEntity);
 
         // Act
         $this->useCase->handle($webHookTransfer);
